突破1

+++ TaFangSSS/Assets/Scripts/Window/英雄详情弹窗.cs
@@ -229,7 +229,7 @@ public class 英雄详情弹窗 : MonoBehaviour
         升级button.onClick.AddListener(() =>
         {
             int 需要值=法则config.法则升级材料Dic[PlayerData.S.英雄法则等级Dic[HeroType]];
-            int 当前值 = PlayerData.S.PropListDic[法则config.法则TypeDic[HeroType]];
+            long 当前值 = PlayerData.S.PropListDic[法则config.法则TypeDic[HeroType]];
             if (当前值 < 需要值)
             {
                 ObserverModuleManager.S.SendEvent("播放音效",音效Type.错误);
